move reflection to Geyser
